Fix duplicate order item steps
AB#7840
AB#7859
